pyproject.toml
[project]
name = "cryptid-flask"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.3",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "sqlalchemy>=2.0.49",
]
